improvement and refactoring
