--- src/chrome/__mocks__/audio.js
@@ -1,13 +1,0 @@
-// mock Audio
-
-export class MockAudio {
-  constructor(src = '') {
-    this.src = src;
-  }
-  play() {
-    this.onended();
-  }
-  onended() {
-    // blank
-  }
-}
