+++ src/chrome/handleWork.js
@@ -1,10 +1,9 @@
 import listenStoreChanges from '../listenStoreChanges';
 import { workFinished } from '../actions';
-import { getAndPlaySoundNotificationOption } from './playSound';
+import { audioPlayer, playSound } from '../playSound';
 
 const handleWork = (store, chrome) => {
   let workTabId = null;
-  const audioPlayer = new window.Audio();
 
   const handleAssignWork = ({ worker: prevWorker }, { worker: curWorker }) => {
     if (prevWorker.get('state') !== 'working' && curWorker.get('state') === 'working') {
@@ -16,7 +15,8 @@ const handleWork = (store, chrome) => {
       const oldWorkTabId = workTabId;
       chrome.tabs.create({ url }, tab => {
         workTabId = tab.id;
-        getAndPlaySoundNotificationOption(store, audioPlayer);
+        const { plugin } = store.getState();
+        playSound(audioPlayer, plugin.get('options'));
       });
       if (oldWorkTabId) {
         chrome.tabs.remove(oldWorkTabId);
